Implemented FileOperations on Unix systems
